Added tests privately distance function still being debugged
