+++ phylo/src/phylo_info/phyloinfo_builder.rs
@@ -9,8 +9,7 @@ use crate::alphabets::Alphabet;
 use crate::io::{self, DataError};
 use crate::parsimony::ParsimonyAligner;
 use crate::phylo_info::PhyloInfo;
-use crate::tree::{Tree};
-use crate::tree::nj_builder::NJBuilder;
+use crate::tree::{nj_builder::NJBuilder, tree_builder::TreeBuilder, Tree};
 use crate::Result;
 
 pub struct PhyloInfoBuilder {
@@ -167,7 +166,7 @@ impl PhyloInfoBuilder {
             None => {
                 info!("Building NJ tree from sequences");
                 let tree_builder = NJBuilder::new(0.00)?;
-                tree_builder.build_nj_tree(&sequences)?
+                tree_builder.build_tree(&sequences)?
                 // This will also work for now
                 //NJBuilder::build_nj_tree(&sequences)?
             }
